Support @Components without @Views as the default case, require template (#23)
In the view building task, check for @Component without a @View and
handle it. Also make sure that the @View or @Component define a template
or templateUrl but not both. Tests included.

Changed most tests that used @View + @Component to just use @Component
since that is the primary use case. Kept two positive test cases for
@View + @Component, adding corresponding ones for simply @Component by
itself. Hopefully those are adequate for ensuring support for @View,
since the code is (at least currently) very similar for both workflows.