Validate inputs and their types in templates - refs #42
Make validating attribute values take the directives attached to the
component. In order to prevent double matching selectors (which, as
written now, would resolve their ranges twice), store them on the
ElementInfo object.

Also rename resolveNodeNames to resolveDirectives because that's what it
does.

Originally I had done some refactoring to not require a mutation of the
ElementInfo object, but due to the nature of #localRefs, it requires
more than one pass over the AST (one to resolve the directives and load
the local vars, and then another to typecheck the expressions with the
scope finalized). Therefore I could either mutate the AST or copy
it/produce a new IR, which seems unwarranted.

Tests; some were updated because they didn't have valid usages of inputs
and not. Others are new to test the new behavior.