Track output fields in the Component and Directive object models #39 #40 #32
Track the outputs in the models and resolve them based on annotations on
the properties or the outputs field on the Component annotation itself
(with or without special rebound names).

Track outputs on standard html components, too.

Some refactoring because inputs and outputs are so similar. Use the
tuple library to break apart 'this: that' since both use that. Keep the
recursive inheritance hierarchy scanner when searching for either inputs
or outputs. More could be done to reduce copying, but differences are
just enough, and more syntactic/structural than semantic (method names
and constructors change) so could be worth it could not be worth it.

A lot of renaming from 'input(s)' to 'inputElement(s)' because 'outputs'
on the task would shadow the property, so went with 'outputElement(s)'
instead and changed 'input(s)' to match.

Resolve those names by tracking AttributeInfo as bound to an input,
property, or none of the above thanks to a new enum (or null).

Tests on that resolution revealed that we were deliberately not tracking
the source range for input names when defined via an empty annotation
(ie `@Input() myprop`). Changed that, because 1. I don't see what the
point of that was and 2. it interfered with my test unless I switched to
the less common outputs: [] format, and 3. it wasn't consistent with
what we were doing with the html components anyway.