Fix issue of template attrs being resolved twice, once without directives (#63)
fixes two tests: https://travis-ci.org/dart-lang/angular_analyzer_plugin/jobs/176588080

This only applies to <template ngFor...>, note that the sugared template
syntaxes are handled differently and don't have this problem because the
sugared attributes are handled before the other attributes are.

No longer resolve template attributes before the template directives are
resolved. Still, we have to resolve the ngFor attribute before we can
put the ngFor variables in scope. So, split that apart from the other
types of vars that are defined based on directives. In the case of
resolving attributes for a template, do the ngFor resolution later but
not too late; in the case of sugared templates, do the same order.