Fix for issue #77 some dom events not recognized
Don't take the dart:html method name and chop off the 'on', that's not
how angular does it. Instead take the @DomName('Document.oneventname')
and get just the eventname out of that.

Previous tests used 'keyUp' as a testcase, but the correct angular is
'keyup'. Note that this difference is in dart:html and was kept in
mock_sdk.dart