Support [attr.x], [class.x], [style.x], and [style.x.px] refs #82 (#88)
* Support [attr.x], [class.x], [style.x], and [style.x.px] refs #82

Validate binding names when they are one of
these special bindings. Check what we can easily check with these
including:

* class.x has a valid classname for x
* class.x is bound to a boolean
* style.x has a valid css identifier for x
* style.x.y has a valid css identifier for x
* style.x.y has a valid css identifier for y (units)
* style.x.y is bound to a number

Went to validate attr.x, but if x is an invalid attr name, then [attr.x]
is an invalid attr name. And there's no typechecking to be done here, at
least, not until we try to discover specific attributes and do stuff
with that (however...even that, I think it just all becomes strings
eventually with .toString(), making there nothing to check)

Had added, but rolled back dart identifier validation in [] () and [()]

Its actually totally valid to name an input/output an invalid dart
identifier and bind to it in your template. Don't prevent that.