Best current solution to handle all types of generic directives refs #86 (#95)
* Best current solution to handle all types of generic directives refs #86

In the future, maybe there's a way to know the directive's parameters
based on how its used in the template. For now, use the upper bounds
because those are the lowest types and therefore can be relied upon in
the template, and avoids the 'T is not subtype of String' errors.

Refactoring: changed event typechecking to go off of a new type field on
InputElement so that it could be deparameterized consistently at
collection time rather than deparameterized everywhere its used. Same
for outputs, but those already had an eventType field that was used
everywhere.

Opening new issue #91 for some tricky subtleties in generic chains.
Looks like this code should some day use `instatiateToBounds` but that
will not do much typing without strong mode turned on, and has a bug
that is open in their repo.

* Review feedback from jmesserly