Preliminary autocompletion work for review
Copy some test code from analysis server to be able to write
autocompletion tests. Basic plugin to autocomplete and dart
autocomplete (which are separate pluggable things).

Store expressions in templates with their offsets so that we can get the
ASTs after the fact for autocomplete.

Instantiate and invoke the some dart completers directly: inheritance
contributor which gathers members based on the enclosing class of the
template, and typed member contributor.

Using these contributors requires a DartCompletionRequest, and the impl
does a lot of things with CompilationUnits which we don't have. So we
implement the class ourselves and shut out the things we don't care
about. Some code duplication here unfortunately.