Create an AST for angular templates and resolve with visitors
The AST is mostly just the ElementInfo and AttributeInfo classes from
before, but with more subclassing involved now (template, text,
expression-bound, and statement-bound attributes split up). Also a node
for mustaches, and the local scope stored on the tree as well for
autocomplete support.

Move the html-ast to angular-ast conversion into converter.dart, which
parser the dart up front, as well as template attributes up front, just
doesn't resolve the dart expressions/statements it builds yet.

Once the conversion is done, break the resolution up into three
visitors:
1. straight shot directive resolver. Since template attributes are
broken down already in the AST, we can resolve directives on the
elemenst and the sugared templates in one straight go.
2. Scope preparer. The parent scope has to be resolved before the next
one can be resolved, so it prepares just one scope. Loads all #vars (as
well as ngFor vars from the template below), and store the corresponding
scopes in the AST so they're ready to be resolved. Resolve outputs so
that $event vars can be loaded in as well.
3. Scope resolver. Once the ASTs are prepared with scopes, resolve them
all. Pretty simple.

Jumping the gun to changing the resolution to using the visitors will
help us transition to the other angular parser which we will one day use
which also relies on an AST and visitors.

Could still use some work, specifically in the complexity of how scopes
are chained. The scope preparer and scope resolver kind of share the
same work there, although it's not immediately obvious how to best
abstract that.

Also made the scope preparer load event variables even though its a very
isolated second pass which sholudn't really be done in the same visitor,
but the code to create variables (which $event vars are) is nontrivial,
and this way it gets to utilize the scoping logic of the scope preparer
which is also, as mentioned before, nontrivial and needing an
abstraction.

That said, I think its a nice step for the code in general and will
serve autocompletion well...or at least, its close to a nice change in
the code. Up to the reviewers to decide that I guess.