Mustache offset tinkering to get completing unclosed mustaches better
When calling parseExpression() on whitespace, the resulting expression's
offset is the end of the whitespace. This makes targeting the expression
in `{{  text` difficult. However, we still want to support some spaces,
for example, `{{myfunc(a, b)`. My hacky solution? Only parse expressions
in unclosed mustaches if they don't start with spaces.

Of course, we don't just not not not parse them, because then there
would be nothing in the AST to target. So we deliberately parse with a
length of 0.

Also make sure that the offset/length of mustaches are accurate even
when they start in the middle of the file, are multiples of them, or are
unclosed.