When a declared (in|out)put is malformed, skip it to prevent crashes refs #183
Used to create a malformed InputElement (or OutputElement) that was
added to the directive and crashed the server when used since it was in
a half complete state.

We could set it to something dynamic, but references won't be traceable
and it should be an error to use an erroneous input since that
erroneous input may have been written by someone else without analysis
and you would have a bad day if you were allowed to use such an input,
discussion in #183 on github.