Autocomplete ast (#179)
* Preliminary autocompletion work for review

Copy some test code from analysis server to be able to write
autocompletion tests. Basic plugin to autocomplete and dart
autocomplete (which are separate pluggable things).

Store expressions in templates with their offsets so that we can get the
ASTs after the fact for autocomplete.

Instantiate and invoke the some dart completers directly: inheritance
contributor which gathers members based on the enclosing class of the
template, and typed member contributor.

Using these contributors requires a DartCompletionRequest, and the impl
does a lot of things with CompilationUnits which we don't have. So we
implement the class ourselves and shut out the things we don't care
about. Some code duplication here unfortunately.

* Autocomplete ngFor expressions, ditch EmbeddedExpression class

* Create an AST for angular templates and resolve with visitors

The AST is mostly just the ElementInfo and AttributeInfo classes from
before, but with more subclassing involved now (template, text,
expression-bound, and statement-bound attributes split up). Also a node
for mustaches, and the local scope stored on the tree as well for
autocomplete support.

Move the html-ast to angular-ast conversion into converter.dart, which
parser the dart up front, as well as template attributes up front, just
doesn't resolve the dart expressions/statements it builds yet.

Once the conversion is done, break the resolution up into three
visitors:
1. straight shot directive resolver. Since template attributes are
broken down already in the AST, we can resolve directives on the
elemenst and the sugared templates in one straight go.
2. Scope preparer. The parent scope has to be resolved before the next
one can be resolved, so it prepares just one scope. Loads all #vars (as
well as ngFor vars from the template below), and store the corresponding
scopes in the AST so they're ready to be resolved. Resolve outputs so
that $event vars can be loaded in as well.
3. Scope resolver. Once the ASTs are prepared with scopes, resolve them
all. Pretty simple.

Jumping the gun to changing the resolution to using the visitors will
help us transition to the other angular parser which we will one day use
which also relies on an AST and visitors.

Could still use some work, specifically in the complexity of how scopes
are chained. The scope preparer and scope resolver kind of share the
same work there, although it's not immediately obvious how to best
abstract that.

Also made the scope preparer load event variables even though its a very
isolated second pass which sholudn't really be done in the same visitor,
but the code to create variables (which $event vars are) is nontrivial,
and this way it gets to utilize the scoping logic of the scope preparer
which is also, as mentioned before, nontrivial and needing an
abstraction.

That said, I think its a nice step for the code in general and will
serve autocompletion well...or at least, its close to a nice change in
the code. Up to the reviewers to decide that I guess.

* AST offset fixes, cleanup, autocomplete local vars with tests

* Mustache offset tinkering to get completing unclosed mustaches better

When calling parseExpression() on whitespace, the resulting expression's
offset is the end of the whitespace. This makes targeting the expression
in `{{  text` difficult. However, we still want to support some spaces,
for example, `{{myfunc(a, b)`. My hacky solution? Only parse expressions
in unclosed mustaches if they don't start with spaces.

Of course, we don't just not not not parse them, because then there
would be nothing in the AST to target. So we deliberately parse with a
length of 0.

Also make sure that the offset/length of mustaches are accurate even
when they start in the middle of the file, are multiples of them, or are
unclosed.

* Tests for autocompleting open mustaches, void/nonvoid suggestions

* Formatting and analyzer fixes

* Abstract scopes into its own base visitor, split event scope prep

New base visitor to walk the path to attributes in scope, with special
methods for what it does as it walks over the various types of template
elements etc along the way.

Now prepare scope visitor can just hook into that, for a way better
separation of concerns.

Also allows event prep to be easily split into its own visitor, although
that required creating a DartVariableManager since now two variables
need the semi-complex logic of creating local vars.

Lastly, the previously complex visitElementInfo on AngularScopeResolver
is basically just removed. And bam, all tests pass.

Overall a big win for separation of concerns and reusability and
readability.

* Fix format

* Add completion tests to all tests

* Test fix

* Formt fix

* Remove unneeded old method

* Analysis error fixes

* Autocomplete inputs and outputs..at least, some of them, sometimes.

* Handle empty bindings better

* Support completing inputs in the middle of an attr