Fuzz test cleanup: fewer iterations, fix truncate/add chunk
Was using substring(x) where I intended substring(0, x), leading to
chunks larger than 80 characters of the snippets appearing in the
files, making for lots of clutter that was hard to debug.

Also seems to have a memory leak. Its sad but not terrible. But I
reduced the iterations so that its not as much of a problem.