Fix issue #204 error when '<' is in binding.
Missing resolution step compared to the regular resolution workflow.
Well, technically, we were missingtwo. But one of them: collecting
varibales, is not relevant to us.

This should be a good change for stability in general, though we
probably allow expression types we shouldn't like typed array list
literals that angular won't like.