Fuzz tester generating random dart and html templates (#203)
* Fuzz tester generating random dart and html templates

Also fixed an easy edge case inside lib/src/tasks.dart.

Adds, removes, or reorganizes characters or lines within dart and html
files. Also grabbed random code chunks from the SDK and the
angular2_components library to throw in at random spots.

Already revealed a few bugs, now that its fairly settled I'm going to
leave this running for a good long while and see what it turns up.

* Fuzz test cleanup: fewer iterations, fix truncate/add chunk

Was using substring(x) where I intended substring(0, x), leading to
chunks larger than 80 characters of the snippets appearing in the
files, making for lots of clutter that was hard to debug.

Also seems to have a memory leak. Its sad but not terrible. But I
reduced the iterations so that its not as much of a problem.