Fix issue #199 option value not recognized
Turns out that the option tag, as well as a few others, don't specify
their tagname anywhere in the source of the dartium html file we scan
to generate the standard html components, so it wasn't being recognized
as a component at all.

There's even one component in there that appears at first glance to be
an html component in the same boat, but it isn't. And there's a few
deprecated ones that are like that too. Skip those (testing the
former...the latter...that I care less about).

Just hardcoded a list, and tested that option works just to test that
my list takes effect. Could have tested the rest, but I don't usually
like it when test code and real code match up 1:1 like that would
appear.

Also modify the mock sdk to include the tested cases.