Fix issue #213 crashing *#h1 attribute
Looks like 'value' is emptystr for templateAttributes without a value,
but valueOffset is null. Therefore later code sees the value as
existing but erroneous, and reports an error but the error can't be
hashed since it contains a null.

This is already specifically avoided inside text attributes, expression
attributes, and statement attributes. Now in template attributes too.

Of course, `<h1 *#h1></h1>` seems to cause angular to crash, but it
should actually be valid as the desugared form `<h1 template="#h1"></h1>`
is valid. (and so is the further desugared form `<template #h1><h1>...`).
But we report an error anyways: 'did you mean let?'. So while I don't
want to assert that error or a different one, I do want to assert it
doesn't crash, so that's all the test does even though 'no crash' tests
are usually considered an antipattern/code smell.

Lastly make sure you don't concat 'null' in desugaring the template,
ie, parsing `<h1 template="#h1 null">`. Was empty string before, which
is what we need to keep.