Fix #210 and #205 and probably more -- trying to resolve unfitting ASTs
Not all of the dart AST is acceptable inside expressions in templates,
and some of the unacceptable ones were resulting in errors. We also
weren't reporting errors for types of things that didn't work inside
angular anyway.

Now catch:

* named arguments
* typed map & list literals (previously caused internal crash)
* anonymous function (#205)
* overriding locals (#210)
* overriding members in expressions (angular disallows this)
* symbols
* throw/rethrow
* new
* is & as
* double dot cascade operations
* this/super
* await (tests at least, already was treated as an identifier)

Also keep our pipes solution working which relied on creating
`AsExpression`s which I'm now blocking. Use the obviously synthetic
offset 0 to detect and let that through.

Fixing overriding of locals was messy. Went off of what the resolver
did: passed the element resolver into the left hand side and inspected
the result to see if it was a property or missing, otherwise block it
since that was causing the exception behind #210.