Fix issue #220 *deferred error. Change ngFor autocomplete strategy
Previously I had made ngFor="let x of " try to parse a statement after
"of". Unfortunately, this is not actually a valid parsing requirement
(though it does indicate an issue, that's for a linter to raise, or for
us to raise in tandem with special ngFor detection, because in a general
sense it could be valid).

Also noticed that "*structuralDirective="let x; attr; attr;" would be
reported wrong while messing with the idea of treating *deferred as
*deferred=";" to suppress the error. Left that fix in.

Once we don't report "of" itself as a syntactic error (it isn't as said
above), then we no longer have a partial expression for
DartSnippetExtractor to pick up. So instead, inside DartSnippetExtractor
we look for TextAttributes that could be intended to become
ExpressionBoundAttributes, and create an expression to pass to the dart
completer when that happens.

In the process of doing that, found some issues with findTarget in the
presence of stars. Firstly, name, value, and originalName were getting
all messed up, which I fixed. Secondly, the name length isn't actually
based on the length of the name, as in the case of ngForOf which is
seven characters but only appears as "of" in the template with a length
of two. Therefore I made sure that findTarget goes off of originalName,
and that originalName is correct in the virtual attributes.

Now with text attributes having a separate name/original name, all
attributes do (templates for the possible star, inputs and outputs for
the brackets, and text for the sugar in a star). Therefore move the
orgiinalName back into AttributeInfo from the previous base class
BoundAttributeInfo. But keep that around because of the children
property it has.

Autocompleting dart inside ngFor definitely got a bit funky, but I
think it is doable for now.