Solve issue #223 supply completion replacement length & offset
Added the ideal replacement length & offset assertions in the tests and
went from there.

Create a new visitor to detect what should be replaced in a target, but
also let the dart completion infrastructure override that when
'extractor.dartSnippet' is not null.

Always suggest `<` in tags since we can tell intelliJ to replace those.

Also found an issue where autocompleting `blah="^"` resulted in a
`replacementLength` of 4, forgotten I had treated that as `blah="null"`
in the converter to suppress the "expected identifier" warning since we
had a later warning for the tag being empty already and didn't want to
report two.

So no longer magically insert a null which messes up completion
lengths, but do still suppress that warning using a new class
IgnoringAnalysisErrorListener. Note that I had created a class with the
same name in a different branch to solve #220, when these get merged
those two classes should use the implementation in this branch where
its in its own file.

Not currently working for "<x^<y" because the html parser treats that
as one tag. And not working for "<x" at EOF because that's treated as a
text node. These should be solved by modifying the html parser ideally.