Resolve issue #171 track tags matching complex selectors (#236)
Handle cases like `x[y],[z]`, which doesn't actually match the tag
for `<x z` even though it matches the tag...or maybe I should say,
it matches the element but not because of the tag name.

Do this by returning not true/false, but an enumeration from
selector.match, where the possible values are no match, tag match, or
non tag match.

In the case of or selectors, don't stop on the first non tag match, in
case that a later tag match will occur. In the case of and selectors,
return tag match if any item in the match is a tag match.