Track, validate, and autocomplete transclusions. (#230)
* Track, validate, and autocomplete transclusions.

In order to track transclusions, we need to look into the html
templates of child components before we can check their parents. The
trick there is that this can be cyclic.

But its not really that its cyclic, its really just that its a first
pass. We don't need to know the relationship between the views before
we get the <ng-content>s in the html file for the view. Therefore we've
split VIEWS into VIEWS1 and VIEWS2. VIEWS1 is for getting the minimum
data to build an AST to check for <ng-content> selectors. In VIEWS2 we
wire up the Views that all have ASTs and ngContent tags.

We also have ANGULAR_ASTS as a new target which takes a dart file and
builds the ASTs for each view in that dart file. It is where we parse
inline dart templates, allowing us to consolidate html parsing into a
mixin in src/tasks.dart. It does, however, produce errors for the
templates not the dart file. So that's an output which is lists of
errros mapped to sources.

Then when you resolve a dart or html template, you make sure to depend
upon ANGULAR_ASTS.of(included directives dart files).

New visitor to collect the <ng-content>s, since it requires no
resolution and can be done in a single pass.

In the future, some changes could be made so that when analyzing a
component template, only the child directive & its <ng-content>s are
needed, but the grandchilds aren't (by depending on VIEWS1).

Also create an API for Selectors to produce HTML tag suggestions since
transclusions can be very hard to figure out. Autocompleting them, even
if djanky, is good. Use a constraining/copying system.

Now in order to suggest the proper elements, we need to track the
parents of TextInfos (and ElementInfos too for less obvious reasons).
Track those in converter.dart.

Also, when validating translusions, if a tag is matched because of its
tagname, record that so that the directive resolver doesn't flag it as
unknown.

We still need to handle @ContentChild, I won't know if that comes up
internally until I try this out later.

* Re-add parse errors, remove redundant errorReporter