Autocomplete stars (#310)
* Complete inputs inside of template elements

Complete, for instance, `trackBy` inside of an ngFor.

* Analyze and suggest directives that appear to want `*star` syntax.

If a `TemplateRef` is in the constructor, it works with star syntax.
Warn when not used with star syntax, and warn when star syntax is used
on a directive that does not have this special ctor as otherwise the
template will go unused.

Its not a foolproof system, I don't think. So detect that separately
from detection of ngIf and ngFor without stars. Especially so that in
the cases where this goes wrong, users can opt out of the error without
losing errors for those core directives' validation...and give
different messages.

Not sure what's more likely...that this expects too many things to be
used with stars, or not enough...

Also suggest this in autocomplete for great profit. Assume that these
directives rely on a property binding, so suggest *p for all p which
are property selectors.

Note this is also a less than perfect system, as it includes suggesting
`*ngForOf`, since that's part of `NgFor`s selector. Hardcode that out,
at least for now.

With this plus suggesting `trackBy:`, we now (I think) suggest
everything relevant for two way autocomplete except `template=`
attributes themselves and the `let` keyword. Wanted to do the latter
but it seemed like it might be a bit treacherous: think `ngIf="^"` vs
`ngFor="^"`, it would be the first time we'd suggest both dart and
non-dart things at once.

* Fix analysis errors

* prefer var and final over types

* More lint fixes

* Include colon in suggestion