Fix #309 empty ngIf suggests brackets. Some refactoring of completion
Looks like the angular template parser creates an empty binding, not a
text attr, for things like `*ngIf` and `*ngFor="let item of"`. Match
that.

Then most of the work was about suppressing errors, because
`*ngFor="let item of items"` technically results in an empty `[ngFor]`
binding which neither matches an input nor has an expression. Suppress
that error, without suppressing the error for `*ngIf` without a value,
by checking that the input matches. Other empty bindings like `of`, or
`trackBy` have new checks in `converter.dart`. A misspelled `trackBy`
will still be caught by the old process.

This also had implications for completion of star attrs, which also had
to dance around these virtual `[ngFor]` bindings. I decided to not just
do the minimum change to get this working, but to also refactor
`findTarget` to be part of `DartSnippetExtractor` so that we can
leverage the cleanliness of the visitor pattern over nasty `is` checks
and spaghetti logic. Not that it was that bad. Also renamed it to
`CompletionTargetExtractor`. It can find `target` as well as
`dartSnippet` in one go.