Very very preliminary navigation based on a results stream
Current caveats:
* only dart files emit navigation
* only on a completely fresh calculation (ie, summaries break this)
* only in an untested number of ways

Added a new stream of results, initially started adding a
`fullyResolvedDirectives` array to try to aid things like navigation
and occurrences which may not want to emit nav for alllll data.
However, navigation should likely be split out from this, since we want
to be able to store it in summaries.