Make the tricky empty prefix attrs special and visitable in clean up
By making empty attribute bindings a subclass of TextAttr, with a
method for it directly in visitors, most existing code can be kept
exactly as-is, without having to worry about the attrs.

A few visitor methods have been added to handle where they then matter.

Notably, the code to select targets is much simpler now.