Break circular dependency between server_plugin and analyzer_plugin
Move starter to server_plugin now that it imports server_plugin, so
that its server_plugin importing itself (which is ok) and
analyzer_plugin (which is ok), instead of it being analyzer_plugin
importing itself (which is ok) and server_plugin (which is not ok).

Circular depnedencies are OK for local development, but not publishable
(I don't think?) and not OK with bazel, which matters for shipping this
to googlers internally.

Also clean up bin/server.dart while I'm in there.