Test functions export like classes and accessors; dont export self
While updating exports tests to test functions as well as classes and
accessors, realized that the original class based export tests were
testing a workflow that makes no sense. Add an error, and tests, that
a component exporting itself is an error. Therefore also refactor tests
that used this to exercise regular class exports, to use a different
class which is valid to export.