Handle 'exports' feature -- restrict scope to exported scope & class. (#374)
* Handle 'exports' feature -- restrict scope to exported scope & class.

Looks like previously all accessors, top levels, classes, enums, and
more were available in templates already.

Therefore the feature is more about collecting exports (and validating
how they are defined), and then restricting the available top levels to
what was available.

Looks like the way to restrict the imports requires creating a layer
between the resolver visitors and LibraryElements,
CompilationUnitElements, and ImportElements and Namespaces. Created a
new folder 'facade' for these wrappers. Unfortunately its a lot of
fragile boilerplate, which may doom that part of this implementation to
maintenance and versioning woes in the future. A possible solution to
this is to create an easier-to-maintain interface which resolver can
accept which we can then fulfill based on the exports array, while the
analyzer can fulfill it based on the Element model.

Alternatively, this could be done with mirrors, except that's
deprecated and all that.

Element model solution has some quirks:
- CompilationUnitElements of a LibraryElement's (the definingUnit)
determines the identifiers which don't require importing.
- LibraryElement's imports which carry prefixes
- LibraryElmeent's Namespaces for imported units

A few things are not carefully prevented due to difficulty of
implementation:
- getAncestor may skip over the facades
- elements which are exported may be returned as part of a namespace or
accessor, etc, of a unit or library to which it doesn't actually belong
- going into children may skip over the facades
- enclosingElement may skip over the facades
- LibraryElement.getType(String) may return unexported names
- likely much more because the API is very very very wide

However, as the tests pass, these issue may not be real issues. Likely,
as long as the issues result in suppressing errors rather than false
errors, the priority of any such bugs will be low.

- Abstract _component?.view?.exports ?? []
- catch exports: [MyClass.member]

* Test that redundant errors aren't raised for wrong prefixes

* Autocompletion

* update README support

* Test functions export like classes and accessors; dont export self

While updating exports tests to test functions as well as classes and
accessors, realized that the original class based export tests were
testing a workflow that makes no sense. Add an error, and tests, that
a component exporting itself is an error. Therefore also refactor tests
that used this to exercise regular class exports, to use a different
class which is valid to export.