Opt in to analysis in new plugin arch via yaml file
Since `ServerPlugin` expects a driver and calls methods on it behind
the scenes, return a `NoopDriver` for the analysis roots that don't
have angular enabled.

Our own code has to detect the NoopDrivers and bail out early, but
that's easy enough to add.

Expects your `analysis_options.yaml` to have:

```
angular:
  enabled: true
```

In order to run our analysis.

New dependency on package:yaml, and handle edge cases like file not
actually existing, etc.

Tests for that as well, with mocks thanks to a little FileService that
gets injected.