Disable empty *starAttr validation, with tests (#412)
Our example `*deferredContent` directive has an input, which fails the
test if we add it. Also, change it to `*deferred` for better
consistency.

Adding that aspect to the test, seems we can still pass our test suite
by limiting validation to NgIf. Left a comment to reenable it...likely
we can do this if the input is marked required, in the future, or we
can decide that the usage of `*deferred` in that test *is* invalid, and
change the definition/API so we can reenable this check.