Fix rejection of innerHTML as an input due to missing dart:html annotation (#413)
Fix rejection of innerHTML as an input due to missing dart:html annotation

The public setter (as well as the private member, which realistically
means nothing for most people) should be annotated, but it isn't. As
such, we can't get the DomName() for innerHTML as such (capitalized)
and then apply angular's map to it.

To ensure that we still _suggest_ `innerHtml`, the angular way, while
accepting `innerHTML`, the DOM way, reuse all existing methodologies
after doing a little switcharoo up front.

Luckily this is harmless for in the future when that annotation gets
fixed, but I hear it may be months and months away. Left a todo.