Handle major Angular security contexts; script & style pending. (#421)
* Handle major Angular security contexts; script & style pending.

For an input with a security context, allow binding to that security
context's safe type instead of the input type. For security contexts
with no sanitization available (resource urls ie iframe.src, scripts),
report an error if bound with anything but the safe type.

Style sanitization started but not completed. Looks like we don't allow
binding to [style] at all currently. Need to look into why its not part
of dart:html and what other exceptions we need to handle and how we
need to handle them. Security will likely "just work" then...some new
tests will need to be written and some comments in other tests will
need to be uncommented.

Script sanitization seems to be something entirely different?

Store security context classes on StandardAngular, and then use them to
create StandardHtml. That does mean we need to flip the order that we
get them. I also put in extra blocks to make sure that when our plugin
runs without any package:angular to resolve, it will not try to get
standard angular forever and ever by forcing 'no priority' as a driver.