Fix tests running in checked mode, and make that default on travis
Had one type error, that had to be fixed.

Also was tripping an assertion in the analyzer's autocomplete that
expressions must have parents. The solution here is perhaps too hacky,
but perhaps not. Just wrap in a virtual parenthesized expression.
Doesn't seem like a problem to pass in null for the open and close
paren tokens.