Properly report a hint (not error) for overly complicated templates
Should have been a hint all along, changed that.

Also, seems like I wrote tests for this but just expected the wrong
error. Solved.

Aaaand the implementation is fixed too, that helps. Basically, since
the OffsettingConstantEvaluator doesn't actually know everything
required to actually resolve the string value, some valid string
templates are not actually strings, but rather, "unknown." Don't report
its not a string just because of this.

The fact that we don't get the actual value in this case doesn't
matter; we won't analyze it.