Allow directives: VARIABLE, using a ResolutionStrategy of list/const.
For list values we are able to report offsets of individual items, and
for const values we are not. Resolution is very different in this
respect and APIs used.

Track the angular standard classes which contain directives (view,
component), and use that resolution data to find annotations for the
case of const resolution -- this can be used in other cases too, like
pipes and selectors etc. It can also likely be used for optimizing the
current behavior.

New summary format to differentiate const/list strategy.

Tests that the `directives: VARIABLE` (or other const expression)
strategy works correctly.