Incorporate custom_tag_names into unit signatures for summaries.
No analysis yet dependent on them, but this ensures the analysis won't
be stale once that analysis is added.

Also change 'unknownTagNames' to 'customTagNames' to reflect the
setting name in the yaml file, 'custom_tag_names' which is the better
term for what it does.

Comments about how summaries will be affected by future analysis
options and making sure they are in the right locations/are adequately
differentiable.