Use List instead of QueryList (though still support it, for now).
A couple subtle things:

* deprecate the SHOULD_BE_QUERY_LIST error code but keep it. The cache
includes the message itself, so to fix it we'd have to bust the cache
which causes deep reanalysis. Next time we bust the cache we can drop
it.

* We have no testing infrastructure for different versions of our
angular version, so I can't test the case where QueryList is removed
from the source. However, its used in one place so I think that's OK.
Have a comment to help people avoid that trap.

And some basic things:
* Accept either QueryList or List
* Change all tests to use List (the preferred way now)
* Add one test to ensure QueryList still works and call it good