Test with pub in travis, except for specific branches.
I'll likely keep one branch that's in sync with the SDK at head. To
maintain it, I'll make new PRs to merge in.

For that reason, go off of branch name to decide whether or not to
leave the script functionally working as is (which uses the latest), or
whether to run pub get to avoid having to work against the latest sdk
packages at all times.