Fix crash where we suggest something prefixed by itself, a range error.
Inside `<div *foo=" ... ">`, we look at the `Foo` directive for inputs
`fooBar`, `fooBaz`, and suggest `bar:`, and `baz:`. However, when `Foo`
has a `foo` input, we may try to suggest it at times. In these times,
it crashes because it gets the suffix (empty string) and attempts to
decapitalize the first character (there is no first character), and it
crashes.

We have no business suggesting it. That is to say, we *could* handle
the crash and suggest ':', but that would be a bogus suggestion anyway.

We already do a prefix check, simply a matter of adding an equality
check too.