Allow custom events with custom types to be specified. (#485)
* Allow custom events with custom types to be specified.

```
analyzer:
  plugins:
    angular:
      enabled: true
      custom_events:
        doodle:
          type: DoodleEvent
          path: 'package:doodle/events.dart'
        poodle:
          type: PoodleEvent
          path: 'package:doodle/events.dart'
```

These don't yet correlate to tags, that could be next up, either
allowing a tag name in the custom_events section, or allowing
custom_events within a tag section, or both.

Still haven't fully decoupled [Options] from [FileTracker], but at
least the new hashes are *less* coupled, and I have some ideas to go
the rest of the way cleanly.

Gather the custom events inside the driver, and make it look like a
part of standard html. Track the source of the options file and the
name offsets of the custom events so that it they be supplied during
completion etc.

Currently doesn't re-resolve the custom event type. We'll see how many
people need it to react to change. I have a feeling that won't be
important, since this is mostly used for polymer users, where polymer
won't be under active development any more.

Also no errors reported yet for misconfigured analysis options. That
has to be a new pipeline unfortunately.

* Incorporate feedback from Leon