Correctness/Performance fix: Use FileContentOverlay in navigation
This is a big performance regression in IntelliJ when inside HTML (where
content info isn't as nicely cached for us by the driver). Before
serving completions, for instance, the navigation would be reading in
large files and computing their line infos which was like 90%+ of CPU
time.

By sharing the overlay, navigation is now 0.16% of CPU time and
completions are much snappier.